Established initial testsuite in jdbc-jtsparser module based on TestNG and migrated JtsTestParser class from examples package into the test suite.
